kadai 9

+++ app/Http/Controllers/TasksController.php
@@ -6,6 +6,8 @@ use Illuminate\Http\Request;
 
 use App\Task;
 
+use Illuminate\Support\Facades\Auth;
+
 class TasksController extends Controller
 {
     /**
@@ -15,12 +17,21 @@ class TasksController extends Controller
      */
     public function index()
     {
-        //
-        $tasks = Task::all();
+        $data = [];
+        if(Auth::check()) {
+            $user = \Auth::user();
+            
+            $tasks = $user->tasks()->get();
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+                ];
+        }
         
-        return view('tasks.index', [
-            'tasks' => $tasks,
-         ]);
+        return view('welcome', $data);
+        
+    
     }
 
     /**
@@ -51,10 +62,11 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             'content' => 'required|max:255',
             ]);
-        
+                
         $task = new Task;
         $task ->status = $request->status;
         $task ->content = $request->content;
+        $task ->user_id =Auth::id();
         $task->save();
         
         return redirect('/');
@@ -71,9 +83,16 @@ class TasksController extends Controller
         //
         $task = Task::findOrFail($id);
         
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        //$task->loadRelationsCounts();
+        //$tasks = $task->tasks()->orderBy('created_at', 'desc')->paginatea(10);
+        
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', [
+                'task' => $task,
+            
+            ]);
+        } 
+        return redirect('/');
     }
 
     /**
@@ -85,11 +104,16 @@ class TasksController extends Controller
     public function edit($id)
     {
         //
-        $task = Task::findOrFail($id);
+        $task = \App\Task::findOrFail($id);
         
-        return view('tasks.edit', [
-            'task' => $task,
-         ]);
+        
+            return view('tasks.edit', [
+            'task' => $task,]);
+        
+        
+            return redirect('/');
+        
+        
     }
 
     /**
@@ -124,9 +148,12 @@ class TasksController extends Controller
     public function destroy($id)
     {
         //
-        $task = Task::findOrFail($id);
+        $task = \App\Task::findOrFail($id);
         
-        $task->delete();
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+        
         
         return redirect('/');
     }
